+++ src/test/java/com/tests/LoginPageTest.java
@@ -1,7 +1,6 @@
 package com.tests;
 
 import java.util.List;
-
 import org.testng.Assert;
 import org.testng.annotations.BeforeMethod;
 import org.testng.annotations.Test;
@@ -24,7 +23,7 @@ public class LoginPageTest extends Base
         cp = new CartPage(driver);
     }
 	
-	@Test(enabled = false)
+	@Test
 	public void verifySuccessProductOrderE2E() throws InterruptedException
 	{		
 		lp.enterUserName();
@@ -46,7 +45,7 @@ public class LoginPageTest extends Base
 		Assert.assertEquals(cp.verifySuccessMessage(), "Thank you for your order!");
 	}
 
-	@Test(enabled = false)
+	@Test
 	public void verifyLogoutFunctionality() throws Exception 
 	{
 		lp.enterUserName();
@@ -58,7 +57,7 @@ public class LoginPageTest extends Base
 		lp.enterUserName();
 	}
 	
-	@Test(enabled = false)
+	@Test
 	public void verifyAboutPageFunctionality() throws Exception 
 	{
 		lp.enterUserName();
@@ -73,6 +72,7 @@ public class LoginPageTest extends Base
 		driver.navigate().back();
 	}
 	
+	@Test
 	public void verifyHamburgerMenuOptions(List<String> actual, List<String> expected) 
 	{
 		lp.enterUserName();
