+++ test-output/Suite/Test.xml
@@ -1,0 +1,13 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!-- Generated by org.testng.reporters.JUnitXMLReporter -->
+<testsuite ignored="3" hostname="Rachit" failures="0" tests="0" time="0.025" errors="0" timestamp="2026-01-06T16:15:22 IST">
+  <testcase classname="com.tests.LoginPageTest" name="verifyLogoutFunctionality" time="0.0">
+    <ignored/>
+  </testcase> <!-- verifyLogoutFunctionality -->
+  <testcase classname="com.tests.LoginPageTest" name="verifySuccessProductOrderE2E" time="0.0">
+    <ignored/>
+  </testcase> <!-- verifySuccessProductOrderE2E -->
+  <testcase classname="com.tests.LoginPageTest" name="verifyAboutPageFunctionality" time="0.0">
+    <ignored/>
+  </testcase> <!-- verifyAboutPageFunctionality -->
+</testsuite>
