+++ test-output/junitreports/TEST-com.tests.LoginPageTest.xml
@@ -1,7 +1,44 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!-- Generated by org.testng.reporters.JUnitReportReporter -->
-<testsuite hostname="Rachit" failures="0" tests="3" name="com.tests.LoginPageTest" time="14.201" errors="0" timestamp="2026-01-06T12:46:28 IST" skipped="0">
-  <testcase classname="com.tests.LoginPageTest" name="verifyLogoutFunctionality" time="2.924"/>
-  <testcase classname="com.tests.LoginPageTest" name="verifyAboutPageFunctionality" time="7.052"/>
-  <testcase classname="com.tests.LoginPageTest" name="verifySuccessProductOrderE2E" time="4.225"/>
+<testsuite hostname="Rachit" failures="0" tests="4" name="com.tests.LoginPageTest" time="13.666" errors="1" timestamp="2026-01-06T16:21:06 IST" skipped="0">
+  <testcase classname="com.tests.LoginPageTest" name="verifyLogoutFunctionality" time="2.842"/>
+  <testcase classname="com.tests.LoginPageTest" name="verifyHamburgerMenuOptions" time="2.495">
+    <error message="
+Cannot inject @Test annotated Method [verifyHamburgerMenuOptions] with [interface java.util.List, interface java.util.List].
+For more information on native dependency injection please refer to https://testng.org/#_dependency_injection" type="org.testng.TestNGException">
+      <![CDATA[org.testng.TestNGException: 
+Cannot inject @Test annotated Method [verifyHamburgerMenuOptions] with [interface java.util.List, interface java.util.List].
+For more information on native dependency injection please refer to https://testng.org/#_dependency_injection
+at org.testng.internal.Parameters.checkParameterTypes(Parameters.java:442)
+at org.testng.internal.Parameters.createParametersForMethod(Parameters.java:347)
+at org.testng.internal.Parameters.createParameters(Parameters.java:699)
+at org.testng.internal.Parameters.handleParameters(Parameters.java:885)
+at org.testng.internal.Parameters.handleParameters(Parameters.java:734)
+at org.testng.internal.invokers.ParameterHandler.handleParameters(ParameterHandler.java:72)
+at org.testng.internal.invokers.ParameterHandler.createParameters(ParameterHandler.java:51)
+at org.testng.internal.invokers.TestInvoker$MethodInvocationAgent.invoke(TestInvoker.java:949)
+at org.testng.internal.invokers.TestInvoker.invokeTestMethods(TestInvoker.java:203)
+at org.testng.internal.invokers.TestMethodWorker.invokeTestMethods(TestMethodWorker.java:154)
+at org.testng.internal.invokers.TestMethodWorker.run(TestMethodWorker.java:134)
+at java.base/java.util.ArrayList.forEach(ArrayList.java:1511)
+at org.testng.TestRunner.privateRun(TestRunner.java:741)
+at org.testng.TestRunner.run(TestRunner.java:616)
+at org.testng.SuiteRunner.runTest(SuiteRunner.java:421)
+at org.testng.SuiteRunner.runSequentially(SuiteRunner.java:413)
+at org.testng.SuiteRunner.privateRun(SuiteRunner.java:373)
+at org.testng.SuiteRunner.run(SuiteRunner.java:312)
+at org.testng.SuiteRunnerWorker.runSuite(SuiteRunnerWorker.java:52)
+at org.testng.SuiteRunnerWorker.run(SuiteRunnerWorker.java:95)
+at org.testng.TestNG.runSuitesSequentially(TestNG.java:1274)
+at org.testng.TestNG.runSuitesLocally(TestNG.java:1208)
+at org.testng.TestNG.runSuites(TestNG.java:1112)
+at org.testng.TestNG.run(TestNG.java:1079)
+at org.testng.remote.AbstractRemoteTestNG.run(AbstractRemoteTestNG.java:115)
+at org.testng.remote.RemoteTestNG.initAndRun(RemoteTestNG.java:293)
+at org.testng.remote.RemoteTestNG.main(RemoteTestNG.java:91)
+]]>
+    </error>
+  </testcase> <!-- verifyHamburgerMenuOptions -->
+  <testcase classname="com.tests.LoginPageTest" name="verifyAboutPageFunctionality" time="6.064"/>
+  <testcase classname="com.tests.LoginPageTest" name="verifySuccessProductOrderE2E" time="2.265"/>
 </testsuite> <!-- com.tests.LoginPageTest -->
